Need to redo some state stuff

+++ cmd/npuzzle/heuristic.go
@@ -1,9 +1,9 @@
 package main
 
-func ManhattanDist(state *State) float32 {
+func ManhattanDist(state *State) int {
 	return (0)
 }
 
-func Conflict(state *State) float32 {
+func Conflict(state *State) int {
 	return (0)
 }
